refactor: Move client API functions to subpackage `client`
